+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Don't update Timestamps if they are older then one week
 - Add image zoom for full screen views
 - Copying image now possible from gallery and chat view
+- Removed options to watch inbox and Deltachat folder from advanced settings
 
 ### Fixed
 - Fix overflow in long links inside quotes @naomiceron #2467